requirements.txt
aiogram>=3.0.0
python-dotenv==1.0.0
aiohttp>=3.8.5
flask==2.0.1
python-telegram-bot==20.7
requests==2.31.0
werkzeug==2.0.1
pyOpenSSL==23.3.0
cryptography>=41.0.0
psutil>=5.9.0
deepseek-ai>=0.1.0
openai>=1.0.0
